# backend/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: SalesMate Backend API
  version: 0.2.0
  description: SalesMate Gmail understanding contract, company workspace and explicit
    rule placeholder.
paths:
...
  /api/v1/agent/chat/requests/{request_id}/:
    get:
      operationId: agent_chat_request_status
      parameters:
      - in: path
        name: request_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      security:
      - agentCredential: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
...
  /api/v1/agent/chat/tool-reads/:
    post:
      operationId: agent_chat_tool_read
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                request_id:
                  type: string
                  format: uuid
                name:
                  type: string
                  minLength: 1
                  maxLength: 120
                arguments:
                  type: object
              required:
              - request_id
              - name
              - arguments
              additionalProperties: false
      security:
      - agentCredential: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
        '400':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
        '403':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
        '404':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
        '409':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
        '500':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/agent/chat/tools/:
    get:
      operationId: agent_chat_tools_catalog
      parameters:
      - in: query
        name: page
        schema:
          type: integer
      - in: query
        name: page_size
        schema:
          type: integer
      - in: query
        name: request_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      security:
      - agentCredential: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
...
components:
...
  securitySchemes:
    agentCredential:
      type: apiKey
      in: header
      name: Authorization
      description: Agent <service-token>; never a Gmail token.
    businessToolCredential:
      type: apiKey
      in: header
      name: Authorization
      description: Tool <user-scoped-token>
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    laboratoryIdentity:
      type: apiKey
      in: header
      name: X-Lab-User
      description: Optional laboratory owner username, available only when LAB_OPEN_ACCESS=true; no token required. Omitting it uses the default laboratory identity. Unavailable in normal mode.